refactor: extract digest_worker singletons, hooks, and shared pipeline
- Move 7 singleton getters to digest_singletons.py (134 lines)
- Move 6 post-exchange hooks to post_exchange_hooks.py (349 lines)
- Consolidate cron/drift pipelines into _run_response_pipeline() helper
- Inline route_and_generate() into its only caller (process_tool_dialog)
- Clean unused imports (ConfigService, OrchestratorService, SessionService, etc.)
- digest_worker.py: 2296 → 1800 lines (-22%)
- Evals: ruff 15 (unchanged), vulture 2 (down from 5), pytest 2980/2980 pass
